Remove unused getResponseDate() and getResetHeader() header
